Added ftp tls

+++ pmng_server/ftp_server/ftp_server.js
@@ -4,17 +4,31 @@ const privileges = require("../privileges");
 const database_server = require("../database_server");
 const FTPfs = require("./ftp_fs");
 const blackhole = require("bunyan-blackhole");
+const greenlock_manager = require("../https/greenlock_manager");
 
-const ftpServer = new FtpSrv({
-    url: "ftp://0.0.0.0:21",
-    pasv_url: process.env.SERVER_HOST,
-    pasv_min: 21001,
-    pasv_max: 21999,
-    greeting: "Welcome to the Platform Manager storages FTP server!",
-    log: blackhole()
-});
+const ftpHostType = process.env.FTP_HOST_TYPE;
+const ftpListenAddr = process.env["HOST_" + ftpHostType];
 
-function start() {
+function getServerSecureContext() {
+    // reusing https certificates from greenlock
+    return greenlock_manager.getSecureContext(process.env.ROOT_DOMAIN, true).catch(() => {
+        console.warn("Cannot get FTP TLS certificate because it doesn't exist for the main domain yet.");
+        return false;
+    });
+}
+
+let ftpServer;
+async function start() {
+    ftpServer = new FtpSrv({
+        url: "ftp://" + ftpListenAddr + ":21",
+        pasv_url: ftpListenAddr,
+        pasv_min: 21001,
+        pasv_max: 21999,
+        tls: (process.env.ENABLE_HTTPS.toLowerCase() ? await getServerSecureContext() : false),
+        greeting: "Welcome to the Platform Manager storages FTP server!",
+        log: blackhole("FTP")
+    });    
+
     ftpServer.on('login', ({connection, username, password}, resolve, reject) => {
         database_server.findUserByName(username).then((user) => {
             if(user == null) return reject("Bad credentials");
@@ -26,11 +40,14 @@ function start() {
         }).catch(reject);
     });
 
-
-    ftpServer.listen().then(() => {
-        privileges.drop();
-        logger.info("FTP server started.");
-    });
+    if(process.env["HOST_" + ftpHostType].toLowerCase() != "disabled") {
+        ftpServer.listen().then(() => {
+            privileges.drop();
+            logger.info("FTP server started.");
+        });
+    } else {
+        logger.warn("Cannot start FTP server. Listen mode " + ftpHostType + " is disabled.");
+    }
 }
 
 start();
